requirements.txt
streamlit
streamlit-audiorec
openai-whisper
groq
pandas